+++ core/editor/src/communication/message.rs
@@ -1,5 +1,6 @@
 use crate::message_prelude::*;
 use graphite_proc_macros::*;
+use std::hash::{Hash, Hasher};
 
 pub trait AsMessage: TransitiveChild
 where
@@ -12,11 +13,11 @@ where
 }
 
 #[impl_message]
-#[derive(PartialEq, Clone, Debug)]
+#[derive(Clone, Debug)]
 pub enum Message {
 	NoOp,
 	#[child]
-	Document(DocumentMessage),
+	Documents(DocumentsMessage),
 	#[child]
 	Global(GlobalMessage),
 	#[child]
@@ -27,4 +28,24 @@ pub enum Message {
 	InputPreprocessor(InputPreprocessorMessage),
 	#[child]
 	InputMapper(InputMapperMessage),
+}
+
+impl Hash for Message {
+	fn hash<H>(&self, state: &mut H)
+	where
+		H: Hasher,
+	{
+		unsafe { std::mem::transmute::<&Message, &[u8; std::mem::size_of::<Message>()]>(self) }.hash(state);
+	}
+}
+
+impl PartialEq for Message {
+	fn eq(&self, other: &Message) -> bool {
+		// TODO: Replace with let [s, o] = [self, other].map(|x| unsafe { std::mem::transmute::<&Message, &[u8; std::mem::size_of::<Message>()]>(x) });
+		let vals: Vec<_> = [self, other]
+			.iter()
+			.map(|x| unsafe { std::mem::transmute::<&Message, &[u8; std::mem::size_of::<Message>()]>(x) })
+			.collect();
+		vals[0] == vals[1]
+	}
 }
