+++ core/editor/src/communication/mod.rs
@@ -1,3 +1,6 @@
+use std::collections::hash_map::DefaultHasher;
+use std::hash::{Hash, Hasher};
+
 pub mod dispatcher;
 pub mod message;
 use crate::message_prelude::*;
@@ -20,4 +23,14 @@ where
 	/// Return true if the Action is consumed.
 	fn process_action(&mut self, action: A, data: T, responses: &mut VecDeque<Message>);
 	fn actions(&self) -> ActionList;
+}
+
+pub fn generate_hash<'a>(messages: impl IntoIterator<Item = &'a Message>, ipp: &InputPreprocessor, document_hash: u64) -> u64 {
+	let mut s = DefaultHasher::new();
+	document_hash.hash(&mut s);
+	ipp.hash(&mut s);
+	for message in messages {
+		message.hash(&mut s);
+	}
+	s.finish()
 }
