+++ core/editor/src/document/mod.rs
@@ -1,8 +1,14 @@
 mod document_file;
 mod document_message_handler;
+mod layer_panel;
+mod movement_handler;
 
 #[doc(inline)]
-pub use document_file::{Document, LayerData};
+pub use document_file::LayerData;
 
 #[doc(inline)]
-pub use document_message_handler::{AlignAggregate, AlignAxis, DocumentMessage, DocumentMessageDiscriminant, DocumentMessageHandler, FlipAxis};
+pub use document_file::{AlignAggregate, AlignAxis, DocumentMessage, DocumentMessageDiscriminant, DocumentMessageHandler, FlipAxis};
+#[doc(inline)]
+pub use document_message_handler::{DocumentsMessage, DocumentsMessageDiscriminant, DocumentsMessageHandler};
+#[doc(inline)]
+pub use movement_handler::{MovementMessage, MovementMessageDiscriminant};
