+++ core/editor/src/tool/mod.rs
@@ -2,7 +2,7 @@ pub mod tool_message_handler;
 pub mod tool_options;
 pub mod tools;
 
-use crate::document::Document;
+use crate::document::DocumentMessageHandler;
 use crate::input::InputPreprocessor;
 use crate::message_prelude::*;
 use crate::{
@@ -25,12 +25,20 @@ pub mod tool_messages {
 	pub use super::tools::rectangle::{RectangleMessage, RectangleMessageDiscriminant};
 }
 
-pub type ToolActionHandlerData<'a> = (&'a Document, &'a DocumentToolData, &'a InputPreprocessor);
+pub type ToolActionHandlerData<'a> = (&'a DocumentMessageHandler, &'a DocumentToolData, &'a InputPreprocessor);
 
 pub trait Fsm {
 	type ToolData;
 
-	fn transition(self, message: ToolMessage, document: &Document, tool_data: &DocumentToolData, data: &mut Self::ToolData, input: &InputPreprocessor, messages: &mut VecDeque<Message>) -> Self;
+	fn transition(
+		self,
+		message: ToolMessage,
+		document: &DocumentMessageHandler,
+		tool_data: &DocumentToolData,
+		data: &mut Self::ToolData,
+		input: &InputPreprocessor,
+		messages: &mut VecDeque<Message>,
+	) -> Self;
 }
 
 #[derive(Debug, Clone)]
